app/requirements.txt
pyjwt
redis
requests
starlette
uvicorn
